+++ R/zzz.R
@@ -1,0 +1,3 @@
+# Package environment for storing state
+# Currently used in error reporting scenarios like `last_check_errors`
+schematic_pkgenv <- new.env(parent = emptyenv())